add check if file exist in buildApp script

+++ chardev_example/buildRun.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 echo "Building..."
-rm simple_app.app
+$appFile = "simple_app.app"
+
+
+if [-f "$appFile"]; then
+    rm simple_app.app
+fi
 gcc simple_app.c -o simple_app.app
 echo "Build done..."
 echo "Run simple_app.app...\n"
